+++ src/arden/MainClass.java
@@ -53,6 +53,7 @@ import arden.compiler.Compiler;
 import arden.compiler.CompilerException;
 import arden.constants.ConstantParser;
 import arden.constants.ConstantParserException;
+import arden.engine.EvokeEngine;
 import arden.runtime.ArdenValue;
 import arden.runtime.BaseExecutionContext;
 import arden.runtime.ExecutionContext;
