+++ src/arden/runtime/BaseExecutionContext.java
@@ -16,11 +16,11 @@ import java.util.List;
 import java.util.Map;
 
 import arden.CommandLineOptions;
-import arden.EvokeEngine;
 import arden.MainClass;
 import arden.compiler.CompiledMlm;
 import arden.compiler.Compiler;
 import arden.compiler.CompilerException;
+import arden.engine.EvokeEngine;
 import arden.runtime.MaintenanceMetadata.Validation;
 import arden.runtime.evoke.CallTrigger;
 import arden.runtime.evoke.Trigger;
